requirements.txt
google-generativeai
jupyter
python-dotenv
llama-cpp-python